irqchip/renesas-irqc: Publish irq_domain when it is ready
The call to irq_domain_add_linear() publishes the irq_domain but the
domain cannot be used by a consumer prior to the call to
irq_alloc_domain_generic_chips() and the extra settings done for the
generic chip itself.

Avoid the race condition by first just creating the irq_domain and then
publishing the domain once it is ready to be used.

Fixes: 99c221df33fb ("irqchip/renesas-irqc: Move over to nested generic chip")
Cc: <stable@vger.kernel.org>
Signed-off-by: Herve Codina <herve.codina@bootlin.com>